unapologetic replicate
